relay_cogs

Module time

Time management for COGS measurements.

COGS are derived from measured amount of work, which usually is done by taking a timed measure. To allow different types of time measures other than wall time, this module exposes a Clock trait.

Structs

Duration

A Duration type to represent a span of time, typically used for system timeouts.

Instant

An instant in time.

StaticClock

A Clock which can be set to any value.

SystemClock

A system Clock which uses the monotonically increasing Instant as its source, measuring wall time.

Traits

Clock

A clock for taking COGS measurements.